server/requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.23.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.9.1
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
Werkzeug==0.15.6
